requirements.txt
fpdf==1.7.2
Pillow==9.5.0
Pygments==2.14.0
streamlit==1.20.0
